% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{munge}
\alias{munge}
\title{munge}
\usage{
munge(data, reps = 10, p_swap = 0.5, local_var = 5,
  as.discrete = NULL, nn_ind = NULL)
}
\arguments{
\item{data}{data frame to use as source for synthetic data generation}

\item{reps}{number of replications of data set required}

\item{p_swap}{Probability of swapping means or values between nearest
neighbors. If a vector is passed, will draw from a uniform distribution
between the minimum and maximum values.}

\item{local_var}{Local variance parameter that controls how. If a vector is
passed, will draw from a uniform distribution between the minimum and
maximum values.}

\item{as.discrete}{vector of columns to force to handle as discrete columns}

\item{nn_ind}{(optional) pre-generated nearest neighbors index to use; if
available, this will avoid the need to do the computationally intensive
nearest neighbors calculation and save time}
}
\description{
Generates synthetic data which mimics joint distributions of original data.
See http://www.niculescu-mizil.org/papers/rtpp364-bucila.rev2.pdf. Algorithm
used in paper has been modified to use square root of hamming distance.
}
