% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_combined_distance.R
\name{nn_combined_distance}
\alias{nn_combined_distance}
\title{nearest neighbors combined distance}
\usage{
nn_combined_distance(continuous_data, discrete_data, batchsize = 2000)
}
\arguments{
\item{continuous_data}{data frame containing all continuous variables}

\item{discrete_data}{data frame containing all discrete variables}

\item{batchsize}{subdivide data into batches of this size in order to speed up
computation. Set to zero to force all data into a single batch (will be
extremely slow for large datasets).}
}
\description{
Determines the nearest neighbors distance between each pair of rows by
computing the Euclidean distance of continuous variables (scaled from 0-1
based on min/max) and the square root of the Hamming distance for discrete
variables.
}
